move method for adding opinion to new controller
