+++ src/main/java/pl/akai/bookcrossing/book/BookRestController.java
@@ -6,15 +6,12 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import pl.akai.bookcrossing.model.Book;
-import pl.akai.bookcrossing.model.Opinion;
-import pl.akai.bookcrossing.opinion.OpinionBean;
 
 @RestController
 @RequiredArgsConstructor
 public class BookRestController {
 
     private final BookBean bookBean;
-    private final OpinionBean opinionBean;
 
     @PostMapping("/book/rent")
     public ResponseEntity<Void> bookRental(@RequestBody Book book) {
@@ -24,12 +21,6 @@ public class BookRestController {
         } else {
             return new ResponseEntity<>(HttpStatus.CONFLICT);
         }
-    }
-
-    @PostMapping("/book/{id}/add-opinion")
-    @ResponseStatus(HttpStatus.CREATED)
-    public void opinionSubmit(@PathVariable int id, @RequestBody Opinion opinion) {
-        opinionBean.insertOpinion(opinion, id);
     }
 
     @PatchMapping(value = "/my-books/change-available", consumes = MediaType.APPLICATION_JSON_VALUE)
@@ -46,7 +37,7 @@ public class BookRestController {
 
     @DeleteMapping("/my-books/decline/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public void bookRentRequestDecline(@PathVariable(value = "id")  Integer requestId) {
+    public void bookRentRequestDecline(@PathVariable(value = "id") Integer requestId) {
         bookBean.deleteBookRentRequestsById(requestId);
     }
 }
